+++ R/zzz.r
@@ -1,6 +1,7 @@
+#' @importFrom stats na.omit
 .onAttach = function(...) {
   if (!interactive()) return()
-  msg = "See https://jumpingrivers.shinyapps.io/benchmarkme/ for a Shiny interface to the benchmark data."
-  
-  packageStartupMessage(paste(strwrap(msg), collapse = "\n"))
+
+  msg = "See https://jumpingrivers.shinyapps.io/benchmarkme/ for a Shiny interface to the benchmark data." # nocov
+  packageStartupMessage(paste(strwrap(msg), collapse = "\n")) # nocov
 }