+++ NEWS.md
@@ -1,3 +1,16 @@
+## Version 0.6.0
+  * Adding parallel benchmarks (thanks to @jknowles)
+  * Since JIT has been introduced, just byte compile the package for ease of comparison.
+
+## Version 0.5.1
+  * Add id_prefix to the upload function
+  * Can now run `benchmark_std` if the package is not attached (thanks to @YvesCR)
+  * Nicer version of `print.bytes` (thanks to @richierocks)
+  * Adding parallel benchmarks (thanks to @jknowles)
+  
+## Version 0.5.0
+  * Bug fix in get_byte_compiler when `cmpfun` was used.
+  
 ## Version 0.4.0
   * Update to shinyapps.io example
   * Moved benchmark description to shinyapps.io